+++ src/cmd/regex_join.rs
@@ -207,13 +207,7 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
         reader
             .into_byte_records()
             .parallel_map_custom(
-                |o| {
-                    if let Some(count) = threads {
-                        o.threads(count)
-                    } else {
-                        o
-                    }
-                },
+                |o| o.threads(threads.unwrap_or_else(num_cpus::get)),
                 move |record| -> CliResult<Vec<csv::ByteRecord>> {
                     let mut record = record?;