+++ src/cmd/scrape.rs
@@ -846,13 +846,7 @@ pub fn run(argv: &[&str]) -> CliResult<()> {
             .into_byte_records()
             .enumerate()
             .parallel_map_custom(
-                |o| {
-                    if let Some(count) = threads {
-                        o.threads(count)
-                    } else {
-                        o
-                    }
-                },
+                |o| o.threads(threads.unwrap_or_else(num_cpus::get)),
                 move |(index, result)| -> CliResult<(csv::ByteRecord, Vec<Vec<DynamicValue>>)> {
                     let record = result?;