+++ CHANGELOG.md
@@ -45,7 +45,7 @@
 * Fixing `xan to md` width alignment.
 * Renaming `xan parallel --shell-preprocessing` short flag to be `-H` because it was being overriden by `-S/--source-column`.
 * Adding missing subcommand completions for `xan parallel`.
-* Fixing `xan parallel` default threads count heuristic.
+* Better default threads count heuristics.
 
 *Quality of Life*